refactor: replace parseArticles hook with getFrontendModule for news slider
