+++ config/services.yaml
@@ -5,8 +5,6 @@ services:
 
   ContaoThemesNet\MateThemeBundle\:
     resource: ../src/
-    exclude:
-      - "../src/Module/ModuleNewsListMate.php"
 
   # Migration
   ContaoThemesNet\MateThemeBundle\Migration\:
